made a few changes to logic
I will be working on more addition this weekend. The button mechanics should now work. One thing that needs to be added is the next screen functionality. I will also be changing the way autonScreen works. This will be by selecting different screen colors depending on team, as well as, selecting a driver screen.